+++ qa/sc_evm_backward_transfer.py
@@ -1,24 +1,29 @@
 #!/usr/bin/env python3
 import logging
 import time
+from _decimal import Decimal
 
 import base58
 from eth_abi import decode
 from eth_utils import add_0x_prefix, encode_hex, event_signature_to_log_topic, remove_0x_prefix
 
 from SidechainTestFramework.account.ac_chain_setup import AccountChainSetup
-from SidechainTestFramework.account.ac_utils import generate_block_and_get_tx_receipt
+from SidechainTestFramework.account.ac_use_smart_contract import SmartContract
+from SidechainTestFramework.account.ac_utils import generate_block_and_get_tx_receipt, format_eoa
 from SidechainTestFramework.account.httpCalls.transaction.allWithdrawRequests import all_withdrawal_requests
+from SidechainTestFramework.account.httpCalls.transaction.createEIP1559Transaction import createEIP1559Transaction
 from SidechainTestFramework.account.httpCalls.transaction.withdrawCoins import withdrawcoins
 from SidechainTestFramework.account.httpCalls.wallet.balance import http_wallet_balance
+from SidechainTestFramework.account.simple_proxy_contract import SimpleProxyContract
 from SidechainTestFramework.account.utils import (computeForgedTxFee,
-                                                  convertZenToZennies, convertZenniesToWei)
+                                                  convertZenToZennies, convertZenniesToWei,
+                                                  WITHDRAWAL_REQ_SMART_CONTRACT_ADDRESS, INTEROPERABILITY_FORK_EPOCH)
 from SidechainTestFramework.sc_forging_util import check_mcreference_presence, check_mcreferencedata_presence
 from SidechainTestFramework.scutil import (
-    generate_next_block, generate_next_blocks
+    generate_next_block, generate_next_blocks, EVM_APP_SLOT_TIME
 )
 from test_framework.util import (
-    assert_equal, assert_false, hex_str_to_bytes,
+    assert_equal, assert_false, hex_str_to_bytes, bytes_to_hex_str, forward_transfer_to_sidechain, assert_true, fail,
 )
 
 """
@@ -52,6 +57,7 @@ Test:
         - reach next withdrawal epoch and verify that certificate for epoch 1 was added to MC mempool
           and then to MC/SC blocks.
         - verify epoch 1 certificate, verify backward transfers list    
+        - interoperability tests
 """
 
 
@@ -77,7 +83,8 @@ def check_withdrawal_event(event, source_addr, dest_addr, amount, exp_epoch):
 class SCEvmBackwardTransfer(AccountChainSetup):
 
     def __init__(self):
-        super().__init__(withdrawalEpochLength=10)
+        super().__init__(block_timestamp_rewind=1500 * EVM_APP_SLOT_TIME * INTEROPERABILITY_FORK_EPOCH,
+                         withdrawalEpochLength=10)
 
     def run_test(self):
         time.sleep(0.1)
@@ -345,6 +352,147 @@ class SCEvmBackwardTransfer(AccountChainSetup):
                      "Second BT amount is wrong.")
 
         assert_equal(we1_certHash, we1_sc_cert["hash"], "Certificate hash is different to the one in MC.")
+
+        #######################################################################################################
+        # Interoperability test with an EVM smart contract calling backward transfer native contract
+        #######################################################################################################
+
+        # Create a new sc address to be used for the interoperability tests
+        evm_address_interop = sc_node.wallet_createPrivateKeySecp256k1()["result"]["proposition"]["address"]
+
+        new_ft_amount_in_zen = Decimal('5.0')
+
+        forward_transfer_to_sidechain(self.sc_nodes_bootstrap_info.sidechain_id,
+                                      mc_node,
+                                      evm_address_interop,
+                                      new_ft_amount_in_zen,
+                                      mc_return_address=mc_node.getnewaddress(),
+                                      generate_block=True)
+
+        generate_next_block(sc_node, "first node")
+
+        # Create and deploy evm proxy contract
+        proxy_contract = SimpleProxyContract(sc_node, evm_address_interop)
+
+        native_contract = SmartContract("WithdrawalRequests")
+
+        # Test before interoperability fork
+        method = "getBackwardTransfers(uint32)"
+        native_input = format_eoa(native_contract.raw_encode_call(method, current_epoch_number))
+        try:
+            proxy_contract.do_static_call(evm_address_interop, 1, WITHDRAWAL_REQ_SMART_CONTRACT_ADDRESS, native_input)
+            fail("Interoperability call should fail before fork point")
+        except RuntimeError as err:
+            print("Expected exception thrown: {}".format(err))
+            # error is raised from API since the address has no balance
+            assert_true("reverted" in str(err))
+
+        # reach the Interoperability fork
+        current_best_epoch = sc_node.block_forgingInfo()["result"]["bestBlockEpochNumber"]
+
+        for i in range(0, INTEROPERABILITY_FORK_EPOCH - current_best_epoch):
+            generate_next_block(sc_node, "first node", force_switch_to_next_epoch=True)
+            self.sc_sync_all()
+
+        res = proxy_contract.do_static_call(evm_address_interop, 1, WITHDRAWAL_REQ_SMART_CONTRACT_ADDRESS,
+                                            native_input)
+
+        # res is (bytes20, uint256)[]. Its ABI encoding in this case is
+        # - first 32 bytes is the offset
+        # - second 32 bytes is array length
+        # - the remaining are the bytes representing the various (bytes20, uint256). Each element is formed of 64 bytes,
+        # 32 for bytes20, 32 for uint256
+        res = res[32:]  # cut offset, don't care in this case
+        num_of_wr = int(bytes_to_hex_str(res[0:32]), 16)
+        assert_equal(2, num_of_wr, "wrong number of backward transfer")
+        res = res[32:]  # cut the array length
+
+        elem_size = 64  # 32 + 32 because each elem is a tuple of bytes20 and uint256
+        list_of_elems = [res[i:i + elem_size] for i in range(0, num_of_wr*elem_size, elem_size)]
+
+        wr_1 = decode([('(bytes20,uint256)')], list_of_elems[0])
+        wr_2 = decode([('(bytes20,uint256)')], list_of_elems[1])
+
+        assert_equal(convertZenniesToWei(sc_bt_amount_in_zennies_1), wr_1[0][1], "Wrong amount in wr")
+        mcDestAddr = bytes_to_hex_str(wr_1[0][0])
+
+        mc_address1_pk = base58.b58decode_check(mc_address1).hex()[4:]
+        assert_equal(mc_address1_pk, mcDestAddr,"Wrong mc address")
+
+        assert_equal(convertZenniesToWei(sc_bt_amount_in_zennies_2), wr_2[0][1], "Wrong amount in wr")
+
+        mcDestAddr = bytes_to_hex_str(wr_2[0][0])
+        mc_address2_pk = base58.b58decode_check(mc_address2).hex()[4:]
+        assert_equal(mc_address2_pk, mcDestAddr, "Wrong mc address")
+
+        # Create a backward transfer
+        # First, evm smart contract needs some zen to withdraw:
+        # 1) create a new sc address
+        # 2) ft some zen to new sc address
+        # 3) send some zen from new sc address to proxy smart contract
+
+        createEIP1559Transaction(sc_node, fromAddress=evm_address_interop, toAddress=format_eoa(
+                                 proxy_contract.contract_address),
+                                 nonce=1, gasLimit=230000, maxPriorityFeePerGas=900000000,
+                                 maxFeePerGas=900000000, value=1000000000000)
+        generate_next_block(sc_node, "first node")
+
+        # Create a transaction requesting a withdrawal request using the proxy smart contract
+        method = "backwardTransfer(bytes20)"
+        native_input = format_eoa(native_contract.raw_encode_call(method, hex_str_to_bytes(mc_address1_pk)))
+
+        bt_amount_in_zennies = 100
+        bt_amount_in_wei = convertZenniesToWei(bt_amount_in_zennies)
+
+        # Estimate gas. The result will be compared with the actual used gas
+        exp_gas = proxy_contract.estimate_gas(evm_address_interop, 2, WITHDRAWAL_REQ_SMART_CONTRACT_ADDRESS,
+                                              bt_amount_in_wei, native_input)
+
+        logging.info("exp_gas: {}".format(exp_gas))
+
+        tx_id = proxy_contract.call_transaction(evm_address_interop, 2, WITHDRAWAL_REQ_SMART_CONTRACT_ADDRESS,
+                                                bt_amount_in_wei, native_input)
+        receipt = generate_block_and_get_tx_receipt(sc_node, tx_id)
+        logging.info("receipt: {}".format(receipt))
+        logging.info("gas used in receipt: {}".format(receipt['result']['gasUsed']))
+
+        # Check the status of tx
+        status = int(receipt['result']['status'], 16)
+        assert_equal(1, status, "Wrong tx status in receipt")
+        # Check the logs
+        assert_equal(1, len(receipt['result']['logs']), "Wrong number of events in receipt")
+        wr_event = receipt['result']['logs'][0]
+        check_withdrawal_event(wr_event, format_eoa(proxy_contract.contract_address), mc_address1, bt_amount_in_zennies, 2)
+
+        # Compare estimated gas with actual used gas. They are not equal because, during the tx execution, more gas than
+        # actually needed is removed from the gas pool and then refunded. This causes the gas estimation algorithm to
+        # overestimate the gas.
+        gas_used = int(receipt['result']['gasUsed'], 16)
+        estimated_gas = int(exp_gas['result'], 16)
+        assert_true(estimated_gas >= gas_used, "Wrong estimated gas")
+
+        # Check tracer
+        trace_response = sc_node.rpc_debug_traceTransaction(tx_id, {"tracer": "callTracer"})
+        logging.info(trace_response)
+
+        assert_false("error" in trace_response)
+        assert_true("result" in trace_response)
+        trace_result = trace_response["result"]
+
+        assert_equal(proxy_contract.contract_address.lower(), trace_result["to"].lower())
+        assert_equal(1, len(trace_result["calls"]))
+        native_call = trace_result["calls"][0]
+        assert_equal("CALL", native_call["type"])
+        assert_equal(proxy_contract.contract_address.lower(), native_call["from"].lower())
+        assert_equal("0x" + WITHDRAWAL_REQ_SMART_CONTRACT_ADDRESS, native_call["to"])
+        assert_true(int(native_call["gas"], 16) > 0)
+        assert_true(int(native_call["gasUsed"], 16) > 0)
+        assert_equal("0x" + native_input, native_call["input"])
+        assert_equal(elem_size * 2 + 2, len(native_call["output"]))  # 130 = 64 bytes * 2 + 0x
+        assert_false("calls" in native_call)
+
+        gas_used_tracer = int(trace_result['gasUsed'], 16)
+        assert_true(gas_used == gas_used_tracer, "Wrong gas")
 
 
 if __name__ == "__main__":