+++ sdk/src/main/scala/io/horizen/account/api/rpc/service/EthService.scala
@@ -28,7 +28,7 @@ import io.horizen.account.wallet.AccountWallet
 import io.horizen.api.http.SidechainTransactionActor.ReceivableMessages.BroadcastTransaction
 import io.horizen.chain.SidechainBlockInfo
 import io.horizen.evm.results.ProofAccountResult
-import io.horizen.evm.{Address, Hash, TraceOptions}
+import io.horizen.evm.{Address, Hash, TraceOptions, Tracer}
 import io.horizen.forge.MainchainSynchronizer
 import io.horizen.network.SyncStatus
 import io.horizen.network.SyncStatusActor.ReceivableMessages.GetSyncStatus
@@ -860,9 +860,6 @@ class EthService(
 
     // get state at previous block
     getStateViewAtTag(nodeView, (blockInfo.height - 1).toString) { (tagStateView, blockContext) =>
-      // enable tracing
-      blockContext.enableTracer(config)
-
       // apply mainchain references
       for (mcBlockRefData <- block.mainchainBlockReferencesData) {
         tagStateView.applyMainchainBlockReferenceData(mcBlockRefData)
@@ -871,18 +868,16 @@ class EthService(
       val gasPool = new GasPool(block.header.gasLimit)
 
       // apply all transaction, collecting traces on the way
-      val evmResults = block.transactions.zipWithIndex.map({ case (tx, i) =>
-        tagStateView.applyTransaction(tx, i, gasPool, blockContext)
-        blockContext.getEvmResult
+      val traces = block.transactions.zipWithIndex.map({ case (tx, i) =>
+        using(new Tracer(config)) { tracer =>
+          blockContext.setTracer(tracer)
+          tagStateView.applyTransaction(tx, i, gasPool, blockContext)
+          tracer.getResult.result
+        }
       })
 
       // return the list of tracer results from the evm
-      val tracerResultList = new ListBuffer[JsonNode]
-      for (evmResult <- evmResults) {
-        if (evmResult != null && evmResult.tracerResult != null)
-          tracerResultList += evmResult.tracerResult
-      }
-      tracerResultList.toList
+      traces.toList
     }
   }
 
@@ -940,18 +935,13 @@ class EthService(
           tagStateView.applyTransaction(tx, i, gasPool, blockContext)
         }
 
-        // enable tracing
-        blockContext.enableTracer(config)
-
-        // apply requested transaction with tracing enabled
-        tagStateView.applyTransaction(requestedTx, previousTransactions.length, gasPool, blockContext)
-
-        // return the tracer result from the evm
-        if (blockContext.getEvmResult != null && blockContext.getEvmResult.tracerResult != null)
-          blockContext.getEvmResult.tracerResult
-        else {
-          logger.warn("Unable to get tracer result from EVM")
-          JsonNodeFactory.instance.objectNode()
+        using(new Tracer(config)) { tracer =>
+          // enable tracing
+          blockContext.setTracer(tracer)
+          // apply requested transaction with tracing enabled
+          tagStateView.applyTransaction(requestedTx, previousTransactions.length, gasPool, blockContext)
+          // return the tracer result
+          tracer.getResult.result
         }
       }
     }
@@ -960,7 +950,6 @@ class EthService(
   @RpcMethod("debug_traceCall")
   @RpcOptionalParameters(1)
   def traceCall(params: TransactionArgs, tag: String, config: TraceOptions): JsonNode = {
-
     applyOnAccountView { nodeView =>
       // get block info
       val blockInfo = getBlockInfoById(nodeView, getBlockIdByHashOrTag(nodeView, tag))
@@ -968,19 +957,15 @@ class EthService(
       // get state at selected block
       getStateViewAtTag(nodeView, if (tag == "pending") "pending" else blockInfo.height.toString) {
         (tagStateView, blockContext) =>
-          // enable tracing
-          blockContext.enableTracer(config)
-
-          // apply requested message with tracing enabled
-          val msg = params.toMessage(blockContext.baseFee, settings.globalRpcGasCap)
-          tagStateView.applyMessage(msg, new GasPool(msg.getGasLimit), blockContext)
-
-          // return the tracer result from the evm
-          if (blockContext.getEvmResult != null && blockContext.getEvmResult.tracerResult != null)
-            blockContext.getEvmResult.tracerResult
-          else {
-            logger.warn("Unable to get tracer result from EVM")
-            JsonNodeFactory.instance.objectNode()
+          using(new Tracer(config)) { tracer =>
+            // enable tracing
+            blockContext.setTracer(tracer)
+            // apply requested message with tracing enabled
+            val msg = params.toMessage(blockContext.baseFee, settings.globalRpcGasCap)
+            Try(tagStateView.applyMessage(msg, new GasPool(msg.getGasLimit), blockContext)) match {
+              case Failure(ex) if !ex.isInstanceOf[ExecutionFailedException] => throw ex
+              case _ => tracer.getResult.result // return the tracer result
+             }
           }
       }
     }